+++ src/language/parser.ts
@@ -1,4 +1,4 @@
-import { EventEmitter } from 'events'
+import { EventEmitter } from 'node:events'
 import lodashDebounce from 'lodash.debounce'
 import type TypedEmitter from 'typed-emitter'
 import {