+++ src/option.ts
@@ -1,9 +1,11 @@
-import { tmpdir } from 'os'
-import path from 'path'
+import { tmpdir } from 'node:os'
+import path from 'node:path'
+import type { Config } from '@marp-team/marp-cli'
 import { MarpOptions } from '@marp-team/marp-core'
 import { Options } from 'markdown-it'
 import { nanoid } from 'nanoid'
-import { TextDocument, Uri, workspace } from 'vscode'
+import { TextDocument, Uri, window, workspace } from 'vscode'
+import openExtensionSettings from './commands/open-extension-settings'
 import themes, { ThemeType } from './themes'
 import {
   marpConfiguration,
@@ -15,6 +17,12 @@ import {
 export interface WorkFile {
   path: string
   cleanup: () => Promise<void>
+}
+
+export interface ConfigForCLI extends Config {
+  vscode: {
+    themeFiles: WorkFile[]
+  }
 }
 
 let cachedPreviewOption: MarpOptions | undefined
@@ -30,8 +38,22 @@ const breaks = (inheritedValue: boolean): boolean => {
   }
 }
 
-const enableHtml = () =>
-  marpConfiguration().get<boolean>('enableHtml') && workspace.isTrusted
+const html = () => {
+  if (workspace.isTrusted) {
+    const htmlConf = marpConfiguration().get<string>('html')
+
+    if (htmlConf === 'all') return { value: true }
+    if (htmlConf === 'off') return { value: false }
+
+    // Legacy configuration compatibility
+    const legacyConf = marpConfiguration().get<boolean>('enableHtml')
+    if (legacyConf) return { value: true, legacy: true }
+
+    return { value: undefined }
+  } else {
+    return { value: false }
+  }
+}
 
 const math = () => {
   const conf = mathTypesettingConfiguration()
@@ -55,10 +77,24 @@ export const marpCoreOptionForPreview = (
   baseOption: Options & MarpOptions,
 ): MarpOptions => {
   if (!cachedPreviewOption) {
+    const htmlOption = html()
+
+    if (htmlOption.legacy) {
+      // Show warning for legacy configuration
+      window
+        .showWarningMessage(
+          'The setting "markdown.marp.enableHtml" is deprecated. Please use "markdown.marp.html" instead. Please review your settings JSON to make silence this warning.',
+          'Open Extension Settings',
+        )
+        .then((selected) => {
+          if (selected) void openExtensionSettings()
+        })
+    }
+
     cachedPreviewOption = {
       container: { tag: 'div', id: '__marp-vscode' },
       slideContainer: { tag: 'div', 'data-marp-vscode-slide-wrapper': '' },
-      html: enableHtml() || undefined,
+      html: htmlOption.value,
       inlineSVG: {
         backdropSelector: false,
       },
@@ -83,11 +119,60 @@ export const marpCoreOptionForCLI = async (
 ) => {
   const confMdPreview = workspace.getConfiguration('markdown.preview', uri)
 
+  let browser = marpConfiguration().get<'auto' | 'chrome' | 'edge' | 'firefox'>(
+    'browser',
+  )
+
+  const browserPath = (() => {
+    const browserPath = marpConfiguration().get<string>('browserPath')
+    if (browserPath) {
+      // If `markdown.marp.browserPath` is `auto`, detect the kind of browser by the binary name
+      if (browser === 'auto') {
+        try {
+          const binaryName = path.basename(browserPath).toLowerCase()
+
+          if (binaryName.includes('firefox') || binaryName.includes('fx'))
+            browser = 'firefox'
+        } catch (e) {
+          console.error(e)
+          console.warn(
+            'Failed to detect the kind of browser by the binary name.',
+          )
+        }
+      }
+
+      return browserPath
+    }
+
+    // Legacy compatibility
+    const chromePath = marpConfiguration().get<string>('chromePath')
+    if (chromePath) {
+      // Show warning for legacy configuration
+      window
+        .showWarningMessage(
+          'The setting "markdown.marp.chromePath" is deprecated. Please use "markdown.marp.browserPath" instead. Please review your settings JSON to make silence this warning.',
+          'Open Extension Settings',
+        )
+        .then((selected) => {
+          if (selected) void openExtensionSettings()
+        })
+
+      // Force to use Chrome if `markdown.marp.chromePath` is set
+      browser = 'chrome'
+
+      return chromePath
+    }
+
+    return undefined
+  })()
+
   const baseOpts = {
     allowLocalFiles,
     pdfNotes,
     pdfOutlines: pdfOutlines(),
-    html: enableHtml() || undefined,
+    html: html().value,
+    browser: browser || 'auto',
+    browserPath,
     options: {
       markdown: {
         breaks: breaks(!!confMdPreview.get<boolean>('breaks')),
@@ -96,8 +181,7 @@ export const marpCoreOptionForCLI = async (
       math: math(),
     },
     themeSet: [] as string[],
-    vscode: {} as Record<string, any>,
-  }
+  } as ConfigForCLI
 
   const workspaceFolder = workspace.getWorkspaceFolder(uri)
   const parentFolder = uri.scheme === 'file' && path.dirname(uri.fsPath)
@@ -137,7 +221,7 @@ export const marpCoreOptionForCLI = async (
   ).filter((w): w is WorkFile => !!w)
 
   baseOpts.themeSet = themeFiles.map((w) => w.path)
-  baseOpts.vscode.themeFiles = themeFiles
+  baseOpts.vscode = { themeFiles }
 
   return baseOpts
 }